app/requirements.txt
gradio==4.29.0
huggingface-hub==0.23.4
python-dotenv==1.0.0
llama-index==0.10.14
llama-index-embeddings-huggingface==0.2.2
llama-index-llms-huggingface==0.2.4
sentence-transformers==2.7.0
langchain
faiss-cpu
aina-gradio-theme==2.3

langchain-community==0.2.1 
langchain-core==0.2.1
